+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 12.1
 ---------------
 - MySQL set CREATED_AT on INSERT in rosariosis_mysql.sql
+- Add Admin Delete Permission in Expenses.php, Incomes.php, Salaries.php, StaffPayments.php, Exceptions.php & Profiles.php
 
 Changes in 12.0.2
 -----------------